+++ src/smartcontracts/codeMetadata.ts
@@ -5,6 +5,7 @@ export class CodeMetadata {
     private upgradeable: boolean;
     private readable: boolean;
     private payable: boolean;
+    private payableBySc: boolean;
 
     /**
      * Creates a metadata object. By default, set the `upgradeable` attribute, and uset all others.
@@ -12,11 +13,13 @@ export class CodeMetadata {
      * @param upgradeable Whether the contract is upgradeable
      * @param readable Whether other contracts can read this contract's data (without calling one of its pure functions)
      * @param payable Whether the contract is payable
+     * @param payableBySc Whether the contract is payable by other smart contracts
      */
-    constructor(upgradeable: boolean = true, readable: boolean = false, payable: boolean = false) {
+    constructor(upgradeable: boolean = true, readable: boolean = false, payable: boolean = false, payableBySc: boolean = false) {
         this.upgradeable = upgradeable;
         this.readable = readable;
         this.payable = payable;
+        this.payableBySc = payableBySc
     }
 
     /**
@@ -41,6 +44,13 @@ export class CodeMetadata {
     }
 
     /**
+     * Adjust the metadata (the `payableBySc` attribute), when preparing the deployment transaction.
+     */
+    togglePayableBySc(value: boolean) {
+        this.payableBySc = value;
+    }
+    
+    /**
      * Converts the metadata to the protocol-friendly representation.
      */
     toBuffer(): Buffer {
@@ -55,6 +65,9 @@ export class CodeMetadata {
         }
         if (this.payable) {
             byteOne |= ByteOne.Payable;
+        }
+        if (this.payableBySc) {
+            byteOne |= ByteOne.PayableBySc;
         }
 
         return Buffer.from([byteZero, byteOne]);
@@ -74,14 +87,16 @@ export class CodeMetadata {
         return {
             upgradeable: this.upgradeable,
             readable: this.readable,
-            payable: this.payable
+            payable: this.payable,
+            payableBySc: this.payableBySc
         };
     }
 
     equals(other: CodeMetadata): boolean {
         return this.upgradeable == other.upgradeable &&
             this.readable == other.readable &&
-            this.payable == other.payable;
+            this.payable == other.payable &&
+            this.payableBySc == other.payableBySc;
     }
 }
 
@@ -93,5 +108,6 @@ enum ByteZero {
 
 enum ByteOne {
     Reserved1 = 1,
-    Payable = 2
+    Payable = 2,
+    PayableBySc = 4
 }